refactor(vize): split check/nuxt into cohesive modules
